added some new parameter and stl files
