+++ pygem/__init__.py
@@ -15,9 +15,7 @@ __all__ = [
 
 from .deformation import Deformation
 from .ffd import FFD
-from .cffd import CFFD
-from .bffd import BFFD
-from .vffd import VFFD
+from .cffd import CFFD,BFFD,VFFD
 from .rbf import RBF
 from .idw import IDW
 from .rbf_factory import RBFFactory
