+++ tests/test_datasets/parameters_test_cffd.prm
@@ -1,0 +1,76 @@
+
+[Box info]
+# This section collects all the properties of the FFD bounding box.
+
+# n control points indicates the number of control points in each direction (x, y, z).
+# For example, to create a 2 x 3 x 2 grid, use the following: n control points: 2, 3, 2
+n control points x: 2
+n control points y: 2
+n control points z: 2
+
+# box length indicates the length of the FFD bounding box along the three canonical directions (x, y, z).
+# It uses the local coordinate system.
+# For example to create a 2 x 1.5 x 3 meters box use the following: box length: 2.0, 1.5, 3.0
+box length x: 1.0
+box length y: 1.0
+box length z: 1.0
+
+# box origin indicates the x, y, and z coordinates of the origin of the FFD bounding box. That is center of
+# rotation of the bounding box. It corresponds to the point coordinates with position [0][0][0].
+# See section "Parameters weights" for more details.
+# For example, if the origin is equal to 0., 0., 0., use the following: box origin: 0., 0., 0.
+box origin x: 0.0
+box origin y: 0.0
+box origin z: 0.0
+
+# rotation angle indicates the rotation angle around the x, y, and z axis of the FFD bounding box in degrees.
+# The rotation is done with respect to the box origin.
+# For example, to rotate the box by 2 deg along the z direction, use the following: rotation angle: 0., 0., 2.
+rotation angle x: 0.0
+rotation angle y: 0.0
+rotation angle z: 0.0
+
+
+[Parameters weights]
+# This section describes the weights of the FFD control points.
+# We adopt the following convention:
+# For example with a 2x2x2 grid of control points we have to fill a 2x2x2 matrix of weights.
+# If a weight is equal to zero you can discard the line since the default is zero.
+#
+# | x index | y index | z index | weight |
+#  --------------------------------------
+# |    0    |    0    |    0    |  1.0   |
+# |    0    |    1    |    1    |  0.0   | --> you can erase this line without effects
+# |    0    |    1    |    0    | -2.1   |
+# |    0    |    0    |    1    |  3.4   |
+
+# parameter x collects the displacements along x, normalized with the box length x.
+parameter x: 0   0   0   0.0
+             0   0   1   0.0
+             0   1   0   0.0
+             0   1   1   0.5
+             1   0   0   0.1
+             1   0   1   0.3
+             1   1   0   0.0
+             1   1   1   0.2
+
+
+# parameter y collects the displacements along y, normalized with the box length y.
+parameter y: 0   0   0   0.0
+             0   0   1   0.0
+             0   1   0   0.0
+             0   1   1   0.1
+             1   0   0   0.3
+             1   0   1   0.3
+             1   1   0   0.0
+             1   1   1   0.2
+
+# parameter z collects the displacements along z, normalized with the box length z.
+parameter z: 0   0   0   0.1
+             0   0   1   0.0
+             0   1   0   0.0
+             0   1   1   0.5
+             1   0   0   0.3
+             1   0   1   0.2
+             1   1   0   0.0
+             1   1   1   0.1
